Add rectangle abstraction for pixels and `Square`s
The algorithm for finding the intersectional area between rectangles
is well known. So, this modlue was added to provide an interface
that represents pixels and `Square`s as generic rectangles.

Then, the solution to this problem can be rephrased as

"Calculate the ratio of intersectional area to total pixel area"